Abstract out pre & post processing functions
